Email/requirements.txt
django
djangorestframework
celery
psycopg2